Add partial implementation of cqrs-es.
Adding a partial implementation o cqrs-es, with the option to use custom
de-/serializers.

Previously, we only had DomainEvent and SerializedEvent here, now we are
adding CqrsFramework, Aggregate, and EventStore, with the
PersistedEventStore and MemStore (for testing) implementations.

The PersistedEventStore does not implement everything from cqrs-es, but
adds `Binarize`, which allows us to plug in different de-/serializers.

We also added the test framework from cqrs-es.